+++ src/app/core/adapter/authentication-firebase.service.ts
@@ -2,7 +2,8 @@ import { Injectable, inject } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { environment } from '@env/environment';
 import { catchError, map, Observable, of, throwError } from 'rxjs';
-import { AuthenticationService, EmailAlreadyTakenError, LoginResponse, RegisterPayload } from '../port/authentication.service';
+import { AuthenticationService, LoginResponse, RegisterPayload } from '../port/authentication.service';
+import { EmailAlreadyTakenError } from 'src/app/visitor/signup/domain/email-already-taken.error';
 
 
 /**
